Для player не создавался board мы исправили это 2 + Поменяли тип данных на BIGINT для Id

+++ src/main/java/game/Happy_Zombie_Farm/dto/PlayerBuildingDto.java
@@ -3,7 +3,7 @@ package game.Happy_Zombie_Farm.dto;
 import java.time.OffsetDateTime;
 
 public record PlayerBuildingDto (
-    Integer playerId,
+    Long playerId,
     String buildingCode,
     Short originX,
     Short originY,
